/ ‘cs_3a/lab/18_matrices/matrix.h’

+++ cs_3a/lab/18_matrices/matrix.h
@@ -83,7 +83,7 @@ public:
    */
    double& operator()(int i, int j)
    {
-	   return i * COLUMNS + j;
+	   return static_cast<double>(i * COLUMNS + j);
    }
 
    /**
